P10 app/requirements.txt
adal==1.2.7
aiohttp==3.6.2
aiounittest==1.3.0
applicationinsights==0.11.9
async-timeout==3.0.1
attrs==22.1.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.28
Babel==2.7.0
botbuilder-ai==4.13.0
botbuilder-applicationinsights==4.13.0
botbuilder-core==4.13.0
botbuilder-dialogs==4.13.0
botbuilder-integration-applicationinsights-aiohttp==4.13.0
botbuilder-schema==4.13.0
botframework-connector==4.13.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
click==8.1.3
colorama==0.4.5
cryptography==3.4.8
datatypes-date-time==1.0.0a2
datedelta==1.4
Django==2.2.6
djangorestframework==3.10.3
emoji==1.7.0
Flask==1.1.1
grapheme==0.6.0
idna==2.10
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==1.4.2
MarkupSafe==2.1.1
msal==1.6.0
msrest==0.6.10
msrestazure==0.6.4
multidict==4.7.6
multipledispatch==0.6.0
oauthlib==3.2.1
opencensus-ext-azure
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.4
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.23.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.25.11
Werkzeug==2.2.2
yarl==1.8.1
